+++ tests/util_spec.rs
@@ -1,6 +1,6 @@
 use std::fmt::Debug;
 
-use beancount_parser_2::{BeancountFile, Flag};
+use beancount_parser_2::{parse, BeancountFile, Flag};
 use rstest::rstest;
 
 fn is_normal<T: Sized + Send + Sync + Unpin>() {}
@@ -25,4 +25,12 @@ fn default_flag_is_completed() {
 fn can_convert_from_flag_to_char(#[case] flag: Flag, #[case] expected: char) {
     let actual: char = flag.into();
     assert_eq!(actual, expected);
+}
+
+#[test]
+fn error_debug_impl_is_succinct() {
+    let input = "2023-06-11 * Oops\n\n\n\n\n; end comment";
+    let Err(err) = parse::<f64>(input) else { unreachable!("parsing should fail") };
+    let debug = format!("{err:?}");
+    assert!(!debug.contains("; end comment"), "{}", debug);
 }